start using click and move things out in the meanwhile
