+++ commands/auth/__init__.py
@@ -1,0 +1,6 @@
+import click
+
+
+@click.group(help="Authentication options")
+def auth():
+    pass
