+++ commands/playlist/__init__.py
@@ -1,0 +1,6 @@
+import click
+
+
+@click.group(help="Playlist information")
+def playlist():
+    pass
