+++ commands/user_top/__init__.py
@@ -1,0 +1,6 @@
+import click
+
+
+@click.group(help="Current user's top tracks and artists")
+def user_top():
+    pass
